osetrenie pri nacitani/vymazani neexistujuceho obrazku

+++ SportApp/Core/src/main/java/com/so/core/services/document/DocumentService.java
@@ -69,14 +69,18 @@ public class DocumentService {
 
         if (r != null) {
             File file = new File(PATH + r.getPath());
-            if (file.delete()) {
-                resourceRepo.delete(r);
-                LOG.info(file.getName() + " subor je vymazany!");
-            } else {
-                LOG.error("Delete operation is failed.");
-                throw new AppException(HttpStatus.INTERNAL_SERVER_ERROR, "nepodarilo sa vymazat resource");
+            if (file.exists()) {
+                if (file.delete()) {
+                    resourceRepo.delete(r);
+                    LOG.info(file.getName() + " subor je vymazany!");
+                } else {
+                    LOG.error("Delete operation is failed.");
+                    throw new AppException(HttpStatus.INTERNAL_SERVER_ERROR, "nepodarilo sa vymazat resource");
+                }
+            } else{
+                LOG.info("subor {} neexistuje, ale resource bol vymazany",r.getPath());
             }
-        }else{
+        } else {
             LOG.info("resource je null");
         }
     }
@@ -91,6 +95,10 @@ public class DocumentService {
         String filePath = PATH + name;
         try {
             File file = new File(filePath);
+            if(!file.exists()){
+                throw new AppException(HttpStatus.NOT_FOUND,"image:"+name+" nebol najdeny");
+            }
+            
             FileInputStream fis = new FileInputStream(file);
             byte[] fileBytes;
             try (BufferedInputStream inputStream = new BufferedInputStream(fis)) {
